[mqtt.homeassistant] fix newStyleChannels
This fixes a couple of problems. The biggest is that if multiple components
were single channel OR no-name, they would all get mapped to a "groupId" of "",
and the next time the Thing started, all but one of them would disappear unless
new discovery information was received. While fixing that, I fixed two other
issues:
 * Multi-channel components with a blank name were getting all their channels
   dumped into the top level of no group. In the newStyleChannels world, we
   don't care about the component's name anymore - multi-channel components
   get a group, single-channel components do not.
 * Instead of statically saying that certain components are only ever single
   or multi-channel components, determine at runtime if a component only added
   a single channel, and if so re-classify it as a single-channel component,
   moving its channel out of the group (and just naming it as the group was
   named)

Signed-off-by: Cody Cutrer <cody@cutrer.us>