[mqtt.homeassistant] Clear internal data when stopping
When the thing config changes, core will dispose of the handler,
and then immediately initialize the exact same object again. If
we don't clear the internal data, we'll find conflicts, and cause
channels to be renamed unnecessarily. I believe this also fixes an
issue that I'm never able to remove topics from the config, even
though they no longer exist.

Signed-off-by: Cody Cutrer <cody@cutrer.us>